Move web app to own project
Move the web application to its own project.
Remove use of the Playwright global tool.
