+++ .github/workflows/build.yml
@@ -40,11 +40,6 @@ jobs:
     - name: Setup .NET SDK
       uses: actions/setup-dotnet@v1
 
-    - name: Install Playwright CLI
-      shell: pwsh
-      run: |
-        dotnet tool install --global Microsoft.Playwright.CLI
-
     - name: Build, test and publish
       shell: pwsh
       run: ./build.ps1 -Runtime win10-x64
@@ -108,11 +103,6 @@ jobs:
 
     - name: Setup .NET SDK
       uses: actions/setup-dotnet@v1
-
-    - name: Install Playwright CLI
-      shell: pwsh
-      run: |
-        dotnet tool install --global Microsoft.Playwright.CLI
 
     - name: Run end-to-end tests
       shell: pwsh
