+++ .vscode/launch.json
@@ -6,9 +6,9 @@
       "type": "coreclr",
       "request": "launch",
       "preLaunchTask": "build",
-      "program": "${workspaceFolder}/src/AdventOfCode/bin/Debug/net6.0/AdventOfCode.dll",
+      "program": "${workspaceFolder}/src/AdventOfCode.Site/bin/Debug/net6.0/AdventOfCode.Site.dll",
       "args": [],
-      "cwd": "${workspaceFolder}/src/AdventOfCode",
+      "cwd": "${workspaceFolder}/src/AdventOfCode.Site",
       "stopAtEntry": false,
       "internalConsoleOptions": "openOnSessionStart",
       "serverReadyAction": {
