+++ src/AdventOfCode.Site/.eslintrc.js
