+++ src/AdventOfCode.Site/AdventOfCode.Site.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+  <PropertyGroup>
+    <Description>Web app for solving Advent of Code.</Description>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <NoWarn>$(NoWarn);CA2007</NoWarn>
+    <OutputType>Exe</OutputType>
+    <RootNamespace>MartinCostello.AdventOfCode.Site</RootNamespace>
+    <TargetFramework>net6.0</TargetFramework>
+    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
+    <TypeScriptToolsVersion>latest</TypeScriptToolsVersion>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.TypeScript.MSBuild" PrivateAssets="all" />
+    <None Remove="scripts\ts\**\*.ts" />
+    <TypeScriptCompile Include="scripts\ts\**\*.ts" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AdventOfCode\AdventOfCode.csproj" />
+  </ItemGroup>
+  <Target Name="BundleAssets" BeforeTargets="BeforeBuild">
+    <Exec Command="npm install" Condition=" !Exists('$(MSBuildThisFileDirectory)\node_modules') " />
+    <Exec Command="npm run build" Condition=" ('$(BuildingInsideVisualStudio)' != 'true' and '$(GITHUB_ACTIONS)' != 'true') or !Exists('$(MSBuildThisFileDirectory)\wwwroot\static\js\main.js') " />
+  </Target>
+</Project>
