+++ src/AdventOfCode.Site/Pages/Home/Index.cshtml.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) Martin Costello, 2015. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-#pragma warning disable SA1649
-
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace MartinCostello.AdventOfCode.Pages;
+namespace MartinCostello.AdventOfCode.Site.Pages;
 
 /// <summary>
 /// A class representing the model for the index page.
