+++ src/AdventOfCode.Site/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using MartinCostello.AdventOfCode
+@using MartinCostello.AdventOfCode.Site
+@using MartinCostello.AdventOfCode.Site.Pages
