+++ src/AdventOfCode.Site/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿// Copyright (c) Martin Costello, 2015. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+using System.Runtime.InteropServices;
+
+[assembly: Guid("ec13ef2b-d4c6-43c0-9ed0-9b5832b70154")]
