+++ src/AdventOfCode.Site/scripts/ts/models/ProblemDetails.ts
