+++ src/AdventOfCode.Site/scripts/ts/models/Puzzle.ts
