+++ src/AdventOfCode.Site/scripts/ts/models/Solution.ts
