+++ src/AdventOfCode.Site/scripts/ts/view/Solver.ts
