+++ src/AdventOfCode.Site/scripts/ts/view/UIElements.ts
