+++ src/AdventOfCode.Site/wwwroot/favicon.ico
