+++ src/AdventOfCode.Site/wwwroot/favicon.png
