+++ src/AdventOfCode.Site/wwwroot/manifest.webmanifest
