+++ src/AdventOfCode.Site/wwwroot/static/js/main.js
@@ -1,0 +1,2 @@
+!function r(i,o,s){function a(t,e){if(!o[t]){if(!i[t]){var n="function"==typeof require&&require;if(!e&&n)return n(t,!0);if(u)return u(t,!0);throw(n=new Error("Cannot find module '"+t+"'")).code="MODULE_NOT_FOUND",n}n=o[t]={exports:{}},i[t][0].call(n.exports,function(e){return a(i[t][1][e]||e)},n,n.exports,r,i,o,s)}return o[t].exports}for(var u="function"==typeof require&&require,e=0;e<s.length;e++)a(s[e]);return a}({1:[function(e,t,n){t.exports={default:e("core-js/library/fn/object/define-property"),__esModule:!0}},{"core-js/library/fn/object/define-property":5}],2:[function(e,t,n){"use strict";n.__esModule=!0,n.default=function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}},{}],3:[function(e,t,n){"use strict";n.__esModule=!0;var r,i=e("../core-js/object/define-property"),o=(r=i)&&r.__esModule?r:{default:r};function s(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),(0,o.default)(e,r.key,r)}}n.default=function(e,t,n){return t&&s(e.prototype,t),n&&s(e,n),e}},{"../core-js/object/define-property":1}],4:[function(e,t,n){t.exports=e("regenerator-runtime")},{"regenerator-runtime":23}],5:[function(e,t,n){e("../../modules/es6.object.define-property");var r=e("../../modules/_core").Object;t.exports=function(e,t,n){return r.defineProperty(e,t,n)}},{"../../modules/_core":8,"../../modules/es6.object.define-property":22}],6:[function(e,t,n){t.exports=function(e){if("function"!=typeof e)throw TypeError(e+" is not a function!");return e}},{}],7:[function(e,t,n){var r=e("./_is-object");t.exports=function(e){if(!r(e))throw TypeError(e+" is not an object!");return e}},{"./_is-object":18}],8:[function(e,t,n){t=t.exports={version:"2.6.12"};"number"==typeof __e&&(__e=t)},{}],9:[function(e,t,n){var o=e("./_a-function");t.exports=function(r,i,e){if(o(r),void 0===i)return r;switch(e){case 1:return function(e){return r.call(i,e)};case 2:return function(e,t){return r.call(i,e,t)};case 3:return function(e,t,n){return r.call(i,e,t,n)}}return function(){return r.apply(i,arguments)}}},{"./_a-function":6}],10:[function(e,t,n){t.exports=!e("./_fails")(function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})},{"./_fails":13}],11:[function(e,t,n){var r=e("./_is-object"),i=e("./_global").document,o=r(i)&&r(i.createElement);t.exports=function(e){return o?i.createElement(e):{}}},{"./_global":14,"./_is-object":18}],12:[function(e,t,n){var m=e("./_global"),v=e("./_core"),y=e("./_ctx"),g=e("./_hide"),b=e("./_has"),w="prototype",_=function(e,t,n){var r,i,o,s=e&_.F,a=e&_.G,u=e&_.S,l=e&_.P,c=e&_.B,f=e&_.W,h=a?v:v[t]||(v[t]={}),d=h[w],p=a?m:u?m[t]:(m[t]||{})[w];for(r in n=a?t:n)(i=!s&&p&&void 0!==p[r])&&b(h,r)||(o=(i?p:n)[r],h[r]=a&&"function"!=typeof p[r]?n[r]:c&&i?y(o,m):f&&p[r]==o?function(r){function e(e,t,n){if(this instanceof r){switch(arguments.length){case 0:return new r;case 1:return new r(e);case 2:return new r(e,t)}return new r(e,t,n)}return r.apply(this,arguments)}return e[w]=r[w],e}(o):l&&"function"==typeof o?y(Function.call,o):o,l&&((h.virtual||(h.virtual={}))[r]=o,e&_.R&&d&&!d[r]&&g(d,r,o)))};_.F=1,_.G=2,_.S=4,_.P=8,_.B=16,_.W=32,_.U=64,_.R=128,t.exports=_},{"./_core":8,"./_ctx":9,"./_global":14,"./_has":15,"./_hide":16}],13:[function(e,t,n){t.exports=function(e){try{return!!e()}catch(e){return!0}}},{}],14:[function(e,t,n){t=t.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=t)},{}],15:[function(e,t,n){var r={}.hasOwnProperty;t.exports=function(e,t){return r.call(e,t)}},{}],16:[function(e,t,n){var r=e("./_object-dp"),i=e("./_property-desc");t.exports=e("./_descriptors")?function(e,t,n){return r.f(e,t,i(1,n))}:function(e,t,n){return e[t]=n,e}},{"./_descriptors":10,"./_object-dp":19,"./_property-desc":20}],17:[function(e,t,n){t.exports=!e("./_descriptors")&&!e("./_fails")(function(){return 7!=Object.defineProperty(e("./_dom-create")("div"),"a",{get:function(){return 7}}).a})},{"./_descriptors":10,"./_dom-create":11,"./_fails":13}],18:[function(e,t,n){t.exports=function(e){return"object"==typeof e?null!==e:"function"==typeof e}},{}],19:[function(e,t,n){var r=e("./_an-object"),i=e("./_ie8-dom-define"),o=e("./_to-primitive"),s=Object.defineProperty;n.f=e("./_descriptors")?Object.defineProperty:function(e,t,n){if(r(e),t=o(t,!0),r(n),i)try{return s(e,t,n)}catch(e){}if("get"in n||"set"in n)throw TypeError("Accessors not supported!");return"value"in n&&(e[t]=n.value),e}},{"./_an-object":7,"./_descriptors":10,"./_ie8-dom-define":17,"./_to-primitive":21}],20:[function(e,t,n){t.exports=function(e,t){return{enumerable:!(1&e),configurable:!(2&e),writable:!(4&e),value:t}}},{}],21:[function(e,t,n){var i=e("./_is-object");t.exports=function(e,t){if(!i(e))return e;var n,r;if(t&&"function"==typeof(n=e.toString)&&!i(r=n.call(e)))return r;if("function"==typeof(n=e.valueOf)&&!i(r=n.call(e)))return r;if(!t&&"function"==typeof(n=e.toString)&&!i(r=n.call(e)))return r;throw TypeError("Can't convert object to primitive value")}},{"./_is-object":18}],22:[function(e,t,n){var r=e("./_export");r(r.S+r.F*!e("./_descriptors"),"Object",{defineProperty:e("./_object-dp").f})},{"./_descriptors":10,"./_export":12,"./_object-dp":19}],23:[function(e,t,n){var r=function(){return this}()||Function("return this")(),i=r.regeneratorRuntime&&0<=Object.getOwnPropertyNames(r).indexOf("regeneratorRuntime"),o=i&&r.regeneratorRuntime;if(r.regeneratorRuntime=void 0,t.exports=e("./runtime"),i)r.regeneratorRuntime=o;else try{delete r.regeneratorRuntime}catch(e){r.regeneratorRuntime=void 0}},{"./runtime":24}],24:[function(e,k,t){!function(e){"use strict";var u,l,c,f,h,d,t,n=Object.prototype,p=n.hasOwnProperty,r="function"==typeof Symbol?Symbol:{},i=r.iterator||"@@iterator",o=r.asyncIterator||"@@asyncIterator",s=r.toStringTag||"@@toStringTag",r="object"==typeof k,a=e.regeneratorRuntime;function m(e,t,n,r){var i,o,s,a,t=t&&t.prototype instanceof y?t:y,t=Object.create(t.prototype),r=new C(r||[]);return t._invoke=(i=e,o=n,s=r,a=l,function(e,t){if(a===f)throw new Error("Generator is already running");if(a===h){if("throw"===e)throw t;return L()}for(s.method=e,s.arg=t;;){var n=s.delegate;if(n){var r=function e(t,n){var r=t.iterator[n.method];if(r===u){if(n.delegate=null,"throw"===n.method){if(t.iterator.return&&(n.method="return",n.arg=u,e(t,n),"throw"===n.method))return d;n.method="throw",n.arg=new TypeError("The iterator does not provide a 'throw' method")}return d}var r=v(r,t.iterator,n.arg);if("throw"===r.type)return n.method="throw",n.arg=r.arg,n.delegate=null,d;r=r.arg;if(!r)return n.method="throw",n.arg=new TypeError("iterator result is not an object"),n.delegate=null,d;{if(!r.done)return r;n[t.resultName]=r.value,n.next=t.nextLoc,"return"!==n.method&&(n.method="next",n.arg=u)}n.delegate=null;return d}(n,s);if(r){if(r===d)continue;return r}}if("next"===s.method)s.sent=s._sent=s.arg;else if("throw"===s.method){if(a===l)throw a=h,s.arg;s.dispatchException(s.arg)}else"return"===s.method&&s.abrupt("return",s.arg);a=f;r=v(i,o,s);if("normal"===r.type){if(a=s.done?h:c,r.arg!==d)return{value:r.arg,done:s.done}}else"throw"===r.type&&(a=h,s.method="throw",s.arg=r.arg)}}),t}function v(e,t,n){try{return{type:"normal",arg:e.call(t,n)}}catch(e){return{type:"throw",arg:e}}}function y(){}function g(){}function b(){}function w(e){["next","throw","return"].forEach(function(t){e[t]=function(e){return this._invoke(t,e)}})}function _(s){var t;this._invoke=function(n,r){function e(){return new Promise(function(e,t){!function t(e,n,r,i){e=v(s[e],s,n);if("throw"!==e.type){var o=e.arg;return(n=o.value)&&"object"==typeof n&&p.call(n,"__await")?Promise.resolve(n.__await).then(function(e){t("next",e,r,i)},function(e){t("throw",e,r,i)}):Promise.resolve(n).then(function(e){o.value=e,r(o)},i)}i(e.arg)}(n,r,e,t)})}return t=t?t.then(e,e):e()}}function x(e){var t={tryLoc:e[0]};1 in e&&(t.catchLoc=e[1]),2 in e&&(t.finallyLoc=e[2],t.afterLoc=e[3]),this.tryEntries.push(t)}function E(e){var t=e.completion||{};t.type="normal",delete t.arg,e.completion=t}function C(e){this.tryEntries=[{tryLoc:"root"}],e.forEach(x,this),this.reset(!0)}function j(t){if(t){var e=t[i];if(e)return e.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var n=-1,e=function e(){for(;++n<t.length;)if(p.call(t,n))return e.value=t[n],e.done=!1,e;return e.value=u,e.done=!0,e};return e.next=e}}return{next:L}}function L(){return{value:u,done:!0}}a?r&&(k.exports=a):((a=e.regeneratorRuntime=r?k.exports:{}).wrap=m,l="suspendedStart",c="suspendedYield",f="executing",h="completed",d={},(e={})[i]=function(){return this},(r=(r=Object.getPrototypeOf)&&r(r(j([]))))&&r!==n&&p.call(r,i)&&(e=r),t=b.prototype=y.prototype=Object.create(e),(g.prototype=t.constructor=b).constructor=g,b[s]=g.displayName="GeneratorFunction",a.isGeneratorFunction=function(e){e="function"==typeof e&&e.constructor;return!!e&&(e===g||"GeneratorFunction"===(e.displayName||e.name))},a.mark=function(e){return Object.setPrototypeOf?Object.setPrototypeOf(e,b):(e.__proto__=b,s in e||(e[s]="GeneratorFunction")),e.prototype=Object.create(t),e},a.awrap=function(e){return{__await:e}},w(_.prototype),_.prototype[o]=function(){return this},a.AsyncIterator=_,a.async=function(e,t,n,r){var i=new _(m(e,t,n,r));return a.isGeneratorFunction(t)?i:i.next().then(function(e){return e.done?e.value:i.next()})},w(t),t[s]="Generator",t[i]=function(){return this},t.toString=function(){return"[object Generator]"},a.keys=function(n){var e,r=[];for(e in n)r.push(e);return r.reverse(),function e(){for(;r.length;){var t=r.pop();if(t in n)return e.value=t,e.done=!1,e}return e.done=!0,e}},a.values=j,C.prototype={constructor:C,reset:function(e){if(this.prev=0,this.next=0,this.sent=this._sent=u,this.done=!1,this.delegate=null,this.method="next",this.arg=u,this.tryEntries.forEach(E),!e)for(var t in this)"t"===t.charAt(0)&&p.call(this,t)&&!isNaN(+t.slice(1))&&(this[t]=u)},stop:function(){this.done=!0;var e=this.tryEntries[0].completion;if("throw"===e.type)throw e.arg;return this.rval},dispatchException:function(n){if(this.done)throw n;var r=this;function e(e,t){return o.type="throw",o.arg=n,r.next=e,t&&(r.method="next",r.arg=u),!!t}for(var t=this.tryEntries.length-1;0<=t;--t){var i=this.tryEntries[t],o=i.completion;if("root"===i.tryLoc)return e("end");if(i.tryLoc<=this.prev){var s=p.call(i,"catchLoc"),a=p.call(i,"finallyLoc");if(s&&a){if(this.prev<i.catchLoc)return e(i.catchLoc,!0);if(this.prev<i.finallyLoc)return e(i.finallyLoc)}else if(s){if(this.prev<i.catchLoc)return e(i.catchLoc,!0)}else{if(!a)throw new Error("try statement without catch or finally");if(this.prev<i.finallyLoc)return e(i.finallyLoc)}}}},abrupt:function(e,t){for(var n=this.tryEntries.length-1;0<=n;--n){var r=this.tryEntries[n];if(r.tryLoc<=this.prev&&p.call(r,"finallyLoc")&&this.prev<r.finallyLoc){var i=r;break}}var o=(i=i&&("break"===e||"continue"===e)&&i.tryLoc<=t&&t<=i.finallyLoc?null:i)?i.completion:{};return o.type=e,o.arg=t,i?(this.method="next",this.next=i.finallyLoc,d):this.complete(o)},complete:function(e,t){if("throw"===e.type)throw e.arg;return"break"===e.type||"continue"===e.type?this.next=e.arg:"return"===e.type?(this.rval=this.arg=e.arg,this.method="return",this.next="end"):"normal"===e.type&&t&&(this.next=t),d},finish:function(e){for(var t=this.tryEntries.length-1;0<=t;--t){var n=this.tryEntries[t];if(n.finallyLoc===e)return this.complete(n.completion,n.afterLoc),E(n),d}},catch:function(e){for(var t=this.tryEntries.length-1;0<=t;--t){var n=this.tryEntries[t];if(n.tryLoc===e){var r,i=n.completion;return"throw"===i.type&&(r=i.arg,E(n)),r}}throw new Error("illegal catch attempt")},delegateYield:function(e,t,n){return this.delegate={iterator:j(e),resultName:t,nextLoc:n},"next"===this.method&&(this.arg=u),d}})}(function(){return this}()||Function("return this")())},{}],25:[function(e,t,n){"use strict";var s=i(e("babel-runtime/regenerator")),r=i(e("babel-runtime/helpers/classCallCheck")),e=i(e("babel-runtime/helpers/createClass"));function i(e){return e&&e.__esModule?e:{default:e}}var a=function(e,s,a,u){return new(a=a||Promise)(function(n,t){function r(e){try{o(u.next(e))}catch(e){t(e)}}function i(e){try{o(u.throw(e))}catch(e){t(e)}}function o(e){var t;e.done?n(e.value):((t=e.value)instanceof a?t:new a(function(e){e(t)})).then(r,i)}o((u=u.apply(e,s||[])).next())})};Object.defineProperty(n,"__esModule",{value:!0});(n.ApiClient=void 0,e.default)(o,[{key:"getPuzzles",value:function(){return a(this,void 0,void 0,s.default.mark(function e(){var t;return s.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,fetch("/api/puzzles");case 2:if((t=e.sent).ok){e.next=5;break}throw new Error(t.status.toString(10));case 5:return e.next=7,t.json();case 7:return e.abrupt("return",e.sent);case 8:case"end":return e.stop()}},e,this)}))}},{key:"solve",value:function(i,o){return a(this,void 0,void 0,s.default.mark(function e(){var t,n,r;return s.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return t={method:"POST",body:o},e.next=3,fetch(i,t);case 3:return n=e.sent,e.next=6,n.json();case 6:if(r=e.sent,n.ok)return e.abrupt("return",r);e.next=11;break;case 11:return e.abrupt("return",r);case 12:case"end":return e.stop()}},e,this)}))}}]),e=o;function o(){(0,r.default)(this,o)}n.ApiClient=e},{"babel-runtime/helpers/classCallCheck":2,"babel-runtime/helpers/createClass":3,"babel-runtime/regenerator":4}],26:[function(e,t,n){"use strict";Object.defineProperty(n,"__esModule",{value:!0});var r=e("./view/Solver");document.addEventListener("DOMContentLoaded",function(){(new r.Solver).initialize()})},{"./view/Solver":27}],27:[function(e,t,n){"use strict";var s=o(e("babel-runtime/regenerator")),r=o(e("babel-runtime/helpers/classCallCheck")),i=o(e("babel-runtime/helpers/createClass"));function o(e){return e&&e.__esModule?e:{default:e}}var a=function(e,s,a,u){return new(a=a||Promise)(function(n,t){function r(e){try{o(u.next(e))}catch(e){t(e)}}function i(e){try{o(u.throw(e))}catch(e){t(e)}}function o(e){var t;e.done?n(e.value):((t=e.value)instanceof a?t:new a(function(e){e(t)})).then(r,i)}o((u=u.apply(e,s||[])).next())})};Object.defineProperty(n,"__esModule",{value:!0}),n.Solver=void 0;var u=e("../client/ApiClient"),l=e("./UIElements"),i=((0,i.default)(c,[{key:"initialize",value:function(){return a(this,void 0,void 0,s.default.mark(function e(){var r,i,t,o=this;return s.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return this.elements.years.addEventListener("change",function(){o.onYearChanged()}),this.elements.days.addEventListener("change",function(){o.onDayChanged()}),this.elements.solve.addEventListener("click",function(){o.onSolve()}),this.elements.form.addEventListener("submit",function(e){return e.preventDefault(),!1}),e.next=6,this.client.getPuzzles();case 6:t=e.sent,r=(new Date).getFullYear(),i=!1,t.forEach(function(e){var t=e.year.toString(),n=void 0;o.puzzlesByYear.has(t)?n=o.puzzlesByYear.get(t):o.puzzlesByYear.set(t,n=[]),n.push(e),1===n.length&&((n=document.createElement("option")).setAttribute("data-year",t),n.setAttribute("value",t),n.text=t,o.elements.years.appendChild(n),e.year===r&&(n.selected=!0,i=!0))}),i||(t=this.elements.years.lastChild)&&(t.selected=!0),this.elements.years.dispatchEvent(new Event("change"));case 12:case"end":return e.stop()}},e,this)}))}},{key:"onDayChanged",value:function(){var e=this.elements.days.selectedOptions[0];this.elements.arguments.value="",this.elements.inputFile.value="",this.hide(this.elements.error),this.elements.error.textContent="",this.hide(this.elements.solutionContainer),this.elements.solution.innerHTML="",this.elements.timeToSolve.textContent="",this.elements.visualization.innerHTML="",this.elements.form.setAttribute("action",e.getAttribute("data-location"));var t=parseInt(e.getAttribute("data-minimum-arguments"),10);0===t?(this.hide(this.elements.argumentsContainer),this.elements.arguments.setAttribute("placeholder",""),this.elements.arguments.setAttribute("rows","0"),this.elements.arguments.removeAttribute("required")):(this.show(this.elements.argumentsContainer),this.elements.arguments.setAttribute("placeholder","Requires "+t+" argument"+(1===t?"":"s")),this.elements.arguments.setAttribute("rows",t.toString(10)),this.elements.arguments.setAttribute("required","")),"true"===e.getAttribute("data-requires-data")?(this.elements.inputFile.setAttribute("required",""),this.show(this.elements.inputFileContainer)):(this.elements.inputFile.removeAttribute("required"),this.hide(this.elements.inputFileContainer))}},{key:"onYearChanged",value:function(){this.elements.days.innerHTML="";for(var e=this.puzzlesByYear.get(this.elements.years.value),t=0;t<e.length;t++){var n=document.createElement("option"),r=e[t],i=r.day.toString(10);n.setAttribute("data-day",i),n.setAttribute("data-location",r.location),n.setAttribute("data-minimum-arguments",r.minimumArguments.toString(10)),n.setAttribute("data-requires-data",r.requiresData.toString()),n.setAttribute("data-year",r.year.toString(10)),n.setAttribute("value",i),n.text=i+" - "+r.name,this.elements.days.appendChild(n);i=new Date;r.year===i.getFullYear()&&r.day===i.getDate()&&11===i.getMonth()&&(n.selected=!0)}this.elements.days.dispatchEvent(new Event("change"))}},{key:"hide",value:function(e){e.classList.add(c.hideClass)}},{key:"show",value:function(e){e.classList.remove(c.hideClass)}},{key:"onSolve",value:function(){return a(this,void 0,void 0,s.default.mark(function e(){var t,n,r,i,o=this;return s.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:if(this.elements.solve.disabled=!0,this.elements.solveText.innerText="Solving...",this.show(this.elements.spinner),t=this.elements.form.getAttribute("action"),n=new FormData,this.elements.arguments.value&&this.elements.arguments.value.split("\n").forEach(function(e){n.append("arguments",e)}),1!==this.elements.inputFile.files.length){e.next=14;break}r=this.elements.inputFile.files[0],(i=new FileReader).readAsText(r,"UTF-8"),i.onload=function(e){n.append("resource",e.target.result),o.solve(t,n)},n.append("resource",r,"input.txt"),e.next=16;break;case 14:return e.next=16,this.solve(t,n);case 16:case"end":return e.stop()}},e,this)}))}},{key:"solve",value:function(i,o){return a(this,void 0,void 0,s.default.mark(function e(){var t,n,r=this;return s.default.wrap(function(e){for(;;)switch(e.prev=e.next){case 0:return this.hide(this.elements.error),this.hide(this.elements.solutionContainer),this.elements.error.textContent="",this.elements.solution.innerHTML="",this.elements.timeToSolve.textContent="",this.elements.visualization.innerHTML="",e.next=8,this.client.solve(i,o);case 8:n=e.sent,n=t=n,t.status?(this.hide(this.elements.solutionContainer),this.show(this.elements.error),this.elements.error.textContent=t.detail):(this.hide(this.elements.error),this.elements.timeToSolve.textContent=n.timeToSolve.toFixed(2)+" milliseconds",this.elements.solution.innerHTML="",this.elements.solutionHeader.textContent="Advent of Code "+n.year+" Day "+n.day+" Solution",n.solutions.forEach(function(e){var t=document.createElement("li");t.classList.add("text-monospace"),t.textContent=e,r.elements.solution.appendChild(t)}),n.visualizations.forEach(function(e){var t=document.createElement("pre");100<(t.textContent=e).split("\n").length?t.classList.add("pre-scrollable"):t.classList.remove("pre-scrollable");e=document.createElement("code");e.appendChild(t);t=document.createElement("li");t.appendChild(e),r.elements.visualization.appendChild(t)}),this.show(this.elements.solutionContainer)),this.hide(this.elements.spinner),this.elements.solveText.innerText="Solve!",this.elements.solve.disabled=!1;case 15:case"end":return e.stop()}},e,this)}))}}]),c);function c(){(0,r.default)(this,c),this.elements=new l.UIElements,this.client=new u.ApiClient,this.puzzlesByYear=new Map}(n.Solver=i).hideClass="d-none"},{"../client/ApiClient":25,"./UIElements":28,"babel-runtime/helpers/classCallCheck":2,"babel-runtime/helpers/createClass":3,"babel-runtime/regenerator":4}],28:[function(e,t,n){"use strict";var r,i=e("babel-runtime/helpers/classCallCheck"),o=(r=i)&&r.__esModule?r:{default:r};Object.defineProperty(n,"__esModule",{value:!0}),n.UIElements=void 0,n.UIElements=function e(){(0,o.default)(this,e),this.arguments=document.getElementById("arguments"),this.argumentsContainer=document.getElementById("arguments-container"),this.inputFile=document.getElementById("resource"),this.inputFileLabel=document.getElementById("resource-label"),this.inputFileContainer=document.getElementById("file-container"),this.solution=document.getElementById("solution"),this.solutionContainer=document.getElementById("solution-container"),this.solutionHeader=document.getElementById("solution-header"),this.spinner=document.getElementById("spinner"),this.visualization=document.getElementById("visualization"),this.days=document.getElementById("day"),this.error=document.getElementById("error"),this.form=document.getElementById("form"),this.solve=document.getElementById("solve"),this.solveText=document.getElementById("solve-text"),this.timeToSolve=document.getElementById("time-to-solve"),this.years=document.getElementById("year")}},{"babel-runtime/helpers/classCallCheck":2}]},{},[26]);
+//# sourceMappingURL=main.js.map
