+++ src/AdventOfCode/AdventOfCode.csproj
@@ -1,27 +1,17 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <Description>Solutions for Advent of Code.</Description>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <NoWarn>$(NoWarn);CA2007;CA2225;CA5351</NoWarn>
     <OutputType>Exe</OutputType>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
     <RootNamespace>MartinCostello.AdventOfCode</RootNamespace>
     <TargetFramework>net6.0</TargetFramework>
-    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-    <TypeScriptToolsVersion>latest</TypeScriptToolsVersion>
   </PropertyGroup>
   <ItemGroup>
     <EmbeddedResource Include="Input\**\input.txt" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Toolkit.HighPerformance" />
-    <PackageReference Include="Microsoft.TypeScript.MSBuild" PrivateAssets="all" />
     <PackageReference Include="System.Runtime.Experimental" />
-    <None Remove="scripts\ts\**\*.ts" />
-    <TypeScriptCompile Include="scripts\ts\**\*.ts" />
   </ItemGroup>
-  <Target Name="BundleAssets" BeforeTargets="BeforeBuild">
-    <Exec Command="npm install" Condition=" !Exists('$(MSBuildThisFileDirectory)\node_modules') " />
-    <Exec Command="npm run build" Condition=" ('$(BuildingInsideVisualStudio)' != 'true' and '$(GITHUB_ACTIONS)' != 'true') or !Exists('$(MSBuildThisFileDirectory)\wwwroot\static\js\main.js') " />
-  </Target>
 </Project>
