--- src/AdventOfCode/Pages/_ViewImports.cshtml
@@ -1,2 +1,0 @@
-@using MartinCostello.AdventOfCode
-@using MartinCostello.AdventOfCode.Pages
