+++ tests/AdventOfCode.Tests/Api/UITests.cs
@@ -5,7 +5,7 @@ using Microsoft.Playwright;
 
 namespace MartinCostello.AdventOfCode.Api;
 
-public class UITests : IntegrationTest
+public class UITests : IntegrationTest, IAsyncLifetime
 {
     public UITests(HttpServerFixture fixture, ITestOutputHelper outputHelper)
         : base(fixture, outputHelper)
@@ -133,6 +133,14 @@ public class UITests : IntegrationTest
         });
     }
 
+    public Task InitializeAsync()
+    {
+        InstallPlaywright();
+        return Task.CompletedTask;
+    }
+
+    public Task DisposeAsync() => Task.CompletedTask;
+
     private static async Task<string> GetPuzzleInputAsync(string year, string day)
     {
         var type = typeof(Puzzle);
@@ -144,6 +152,16 @@ public class UITests : IntegrationTest
         using var reader = new StreamReader(stream);
 
         return await reader.ReadToEndAsync();
+    }
+
+    private static void InstallPlaywright()
+    {
+        int exitCode = Microsoft.Playwright.Program.Main(new[] { "install" });
+
+        if (exitCode != 0)
+        {
+            throw new InvalidOperationException($"Playwright exited with code {exitCode}.");
+        }
     }
 
     private async Task<PuzzleSolver> LoadApplication(IPage page)
